Configurable hostname length.
DocImpact: Windows and Linux return different things for a call to
socket.gethostname() (either FQDN or not). This method normalizes
the output to make running mixed environments easier.

The big goal here is to not changing the current default behaviour,
whilst allowing deployments to do something different without
having to set a customized flag on every machine. Therefore, the
default value for the new hostname_length flag is "system", which
will do whatever was happening before. If you want FQDNs, then you
should change that flag to "full". For just the host portion of the
name, use "short".

There's a lot of code churn here to handle the fact that we can't
use the value of a flag to help us calculate the value of the flag.

Resolves bug 1055503.

Change-Id: I13a40c7746f084250b3ad7c2e21900dcd8df666f