Cache "qemu-img info" calls.
We ask qemu regularly if the virtual size of a disk image has changed,
yet the only time it actually will is if we resize the image. We can
therefore cache this information.

To achieve this, I've moved the recently added memoize decorator out
of the EC2 API and into utils. Doing that required making it a little
more generic as well.

Resolves bug 1154162.

Change-Id: I133619f4a2a2707a84b254e951d005689b69ba39