Implement a simple local store.
This store is deliberately unstructured. It is up to the user to handle
versioning and so forth. In the use cases I can think of, its just a
list, so shouldn't require much work.

It doesn't make sense to store this data in the central nova DB,
because if you lose the disks on a compute node, this store is
intended for work which would no longer be required.

Change-Id: I4c49a0f3e7e0039649b8eb80718aae67c78e10d8