Cache "qemu-img info" calls.
We ask qemu regularly if the virtual size of a disk image has changed,
yet the only time it actually will is if we resize the image. We can
therefore cache this information.

This is done with a new bounded cache implementatin that sits in
nova.utils for now. If it ends up being used more generically, I will
move it into oslo.

Change-Id: I42e19e8741a20ffa0366040959846423bc7dc10c