Retry failed instance file deletes.
This change adds a new column to the instances table named "reaped".
This column indicates that the instance files on disk have been
deleted for that instance, assuming that the virt driver you are
using supports that feature (which only libvirt does for now).

Then the flag instance_delete_interval specifies how often to
retry deletes. The default is five minutes. A value of zero will
disable retries.

A second column is added to the instances table as well. This column,
named "reap_attempts" tracks the number of times we have tried to
clean up an instance's files. We give up after
maximum_instance_delete_attempts attempts.

This change is done in this way because reaping will become a more
generic concept in the next patch.

DocImpact: the following new flags are added...

 - instance_delete_interval: how often to try to clean up instance
   files

 - maximum_instance_delete_attempts: how many attempts to have before
   giving up

Resolves bug 1040110.

Change-Id: I895869e18f92ed57bd2f8c1683cf9a398d88b680