Handle inject failures for mandatory files correctly
The logic for the inject failure handling was incorrect, in that it
was treating injection of mandatory files as being a non-fatal error.

I've also moved the static class method to being just a normal method
because it was causing me testing errors once I fixed the error
reporting and I can't see a good reason for it to be a class method.

Change-Id: I77f87cddb8634782947cb3aea1953f169ef53cfd
Closes-Bug: 1221985