Implement unix domain socket consoles for libvirt
For this change to work for you, you'll need to either disable
apparmor or alter the libvirt templates to allow access to
/tmp. You also need to change the value of the
console_log_type flag to "domain". Console logs are then passed
from libvirt to nova-compute with a unix domain socket. This
allows us to implement console log file rotation.

This change also makes the way file system elements are shared
with libvirt configurable. You can specify a chmod value for
those elements (the previous implementation), or a group to
chgrp to which is shared by nova-compute and libvirt. If
neither option is specified, nova-compute will refuse to start
if using the libvirt hypervisor driver.

This change also slightly refactors how all console logging in
libvirt works, but there are no functional changes to "file"
style logging.

New flags:
    console_log_type
    max_rotated_consoles

    libvirt_shared_file_mode
    libvirt_shared_group
    libvirt_shared_noop

DocImpact

SecurityImpact: this should reduce the amount of unbounded
growth we see in libvirt console log files.

Closes-Bug: 832507
Closes-Bug: 1227027
Progresses blueprint fix-libvirt-console-logging

This review squashes a series of previous reviews down to one.
Previous change ids are:
    I6d65acf74e6d2734c16ccea243b3110a97e7f792
    I70e0e21a99fe45dcb2ee3733f382e8f1d0302905
    I5d1cb79375a2ed509262f19af93fa2946e1edf3c

Change-Id: I3513c12ed446171cfba819d8513616a70c983877