Forcing a configdrive in cfg should populate DB.
If you force a config drive for all instances, then it should result
in the config_drive column for the instance being set to true in the
instances table. Also, this should only happen on first boot of an
instance, not for things like hard reboots.

This change means we can't just use force_config_drive as short
hand in testing for virt drivers, so there's some related fallout
there.

DocImpact: the force_config_drive flag is only used on first boot
of an instance. We should tell cloud admins about this.

Change-Id: I09107a7f1e9b2e67f0fcd19dce8fd063d69e1632
Closes-Bug: #1303714
Closes-Bug: #1241806