Handle config drives being stored on rbd
rbd is the only example of a currently supported image storage
backend where it makes sense to put the config drive in the
configured storage backend instead of local hypervisor disk. I
don't think this makes sense for LVM, where we would be creating
a LV for a tens of megabytes file, which seems like overkill to
me. The other storage backends use local disk for their data
already.

This use case was covered by the now reverted changes:
  228d0221763b12f11ecbacde4db38b1151f96e31
  0b01e846d40f3b343da9ebe1dae89cca8bc2ac66
  ecce888c469c62374a3cc43e3cede11d8aa1e799

Support this special case by moving the image to rbd once it has
been created in the local instance directory on the hypervisor.

I've tested this change in devstack and it works.

Change-Id: Ia3ca5a18c79d62b71b9c042a612d12dd074b245e
Related-bug: #1369627
Related-bug: #1361840