We no longer need rootwrap.
(Well, except for starting privsep daemons, oh and nova-net,
oh and of course os-brick wants to know what our root helper is
for some reason that is a bit beyond me.)

Yes that's right, we now live in the future and no longer need
the run_as_root infrastructure in nova.utils. I'm sure this breaks
some out of tree drivers, but they've had an entire release to
notice that we're moving in this direction and its pretty easy
for them to fix themselves these days.

Do not merge this patch until we've gotten brave and deleted
nova-net once and for all. You have been warned.

Change-Id: I99c66558938db9beb0bda33d27a8e36b26b8fcac
blueprint: hurrah-for-privsep