Implement simple prometheus metrics for nova.
This is a toe in the water to expose metrics that I think
operational staff need to monitor nova well. There have been
previous attempts to expose some of this data via nova-api
calls, but it hasn't been great and some of the calls they
make are very expensive.

This isn't prometheus specific, its just the case that I care
about. I've tried reasonably hard to do this in a generic
manner which should work for other monitoring systems, but I'd
be interested in feedback.

Change-Id: I36e911ca935af53d6ef4fc2382a1487c950fa619